Rename SearchBox SearchInput
