+++ src/index.js
@@ -10,7 +10,7 @@ export { default as Button } from './components/Button/Button';
 export { default as Checkbox } from './components/Checkbox/Checkbox';
 export { default as Grid } from './components/Grid/Grid';
 export { default as Icon } from './components/Icon/Icon';
-export { default as SearchBox } from './components/SearchBox/SearchBox';
+export { default as SearchInput } from './components/SearchInput/SearchInput';
 export { default as PageNavBar } from './components/PageNavBar/PageNavBar';
 export { default as Radio } from './components/Radio/Radio';
 export { default as RadioGroup } from './components/RadioGroup/RadioGroup';
